refactor: add useSpeechRecognition, useVoiceRecording, useVoiceVisualization
